[QMS-108] Moved all help to common source code to be shared with QMT
* Created a src/common folder
* moved all help related files to be shared by QMS and QMT
