+++ src/common/help/CHelpBrowser.cpp
@@ -16,24 +16,30 @@
 
 **********************************************************************************************/
 
-#include "widgets/CHelpBrowser.h"
+#include "help/CHelpBrowser.h"
 #include <QtHelp>
 
 CHelpBrowser::CHelpBrowser(QHelpEngine *helpEngine, QWidget *parent)
     : QTextBrowser(parent)
     , engine(helpEngine)
 {
-
     connect(engine->contentWidget(), &QHelpContentWidget::linkActivated, this, &CHelpBrowser::setSource);
     connect(engine->indexWidget(), &QHelpIndexWidget::linkActivated, this, &CHelpBrowser::setSource);
     connect(engine->searchEngine()->resultWidget(), &QHelpSearchResultWidget::requestShowLink, this, &CHelpBrowser::setSource);
-
 }
 
+void CHelpBrowser::setSource(const QUrl& url)
+{
+    if(url.scheme().startsWith("http"))
+    {
+        QDesktopServices::openUrl(url);
+        return;
+    }
+    QTextBrowser::setSource(url);
+}
 
 QVariant CHelpBrowser::loadResource(int type, const QUrl &name)
 {
-    qDebug() << type << name;
     if (name.scheme() == "qthelp")
     {
         return QVariant(engine->fileData(name));
