+++ src/common/help/CHelpBrowser.h
@@ -32,6 +32,9 @@ public:
 
     QVariant loadResource (int type, const QUrl& name) override;
 
+public slots:
+    void setSource(const QUrl& url) override;
+
 private:
     QHelpEngine * engine;
 };
