+++ src/common/help/CHelpSearch.h
