+++ src/qmapshack/IMainWindow.ui
@@ -116,8 +116,8 @@
     </property>
     <addaction name="actionAbout"/>
     <addaction name="actionHelp"/>
+    <addaction name="actionWiki"/>
     <addaction name="actionQuickstart"/>
-    <addaction name="actionNewHelp"/>
    </widget>
    <widget class="QMenu" name="menuProject">
     <property name="title">
@@ -462,16 +462,13 @@
     <enum>QAction::AboutRole</enum>
    </property>
   </action>
-  <action name="actionHelp">
+  <action name="actionWiki">
    <property name="icon">
     <iconset resource="resources.qrc">
      <normaloff>:/icons/32x32/Help.png</normaloff>:/icons/32x32/Help.png</iconset>
    </property>
    <property name="text">
-    <string>Help</string>
-   </property>
-   <property name="shortcut">
-    <string>F1</string>
+    <string>Online Wiki</string>
    </property>
    <property name="menuRole">
     <enum>QAction::NoRole</enum>
@@ -1003,13 +1000,16 @@
     <string>Mark the highlighted track with a red border.</string>
    </property>
   </action>
-  <action name="actionNewHelp">
+  <action name="actionHelp">
    <property name="icon">
     <iconset resource="resources.qrc">
      <normaloff>:/icons/32x32/Help.png</normaloff>:/icons/32x32/Help.png</iconset>
    </property>
    <property name="text">
-    <string>New Help</string>
+    <string>Help</string>
+   </property>
+   <property name="shortcut">
+    <string>F1</string>
    </property>
   </action>
  </widget>
